Forgot password (#187)

+++ lam7a/lib/features/authentication/service/authentication_api_service.dart
@@ -3,6 +3,7 @@ import 'package:lam7a/core/constants/server_constant.dart';
 import 'package:lam7a/core/services/api_service.dart';
 import 'package:lam7a/features/authentication/model/authentication_user_credentials_model.dart';
 import 'package:lam7a/features/authentication/model/authentication_user_data_model.dart';
+import 'package:lam7a/features/authentication/ui/view/screens/forgot_password/reset_password.dart';
 import 'package:lam7a/features/authentication/utils/authentication_constants.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 // import 'package:lam7a/features/authentication/model/user_data_model.dart';
@@ -61,7 +62,6 @@ class AuthenticationApiService {
       endpoint: ServerConstant.login,
       data: userCredentials.toJson(),
     );
-    
   }
 
   Future<void> test() async {
@@ -76,18 +76,21 @@ class AuthenticationApiService {
       endpoint: ServerConstant.logout,
     );
     print(res[AuthenticationConstants.status]);
-    return res[AuthenticationConstants.status] == AuthenticationConstants.success;
+    return res[AuthenticationConstants.status] ==
+        AuthenticationConstants.success;
   }
+
   Future<List<dynamic>> getInterests() async {
     Map<String, dynamic> res = await apiService.get(
       endpoint: ServerConstant.getInterests,
     );
     return res['data'];
   }
+
   Future<List<dynamic>> getUsersToFollow([int limit = 10]) async {
     Map<String, dynamic> res = await apiService.get(
       endpoint: ServerConstant.toFollowUsers,
-      queryParameters: {'limit': limit}
+      queryParameters: {'limit': limit},
     );
     return res['data']['users'];
   }
@@ -98,23 +101,54 @@ class AuthenticationApiService {
       data: {'interestIds': interestIds},
     );
   }
+
   Future<Map<String, dynamic>> followUsers(int userId) async {
-    return await apiService.post(
-      endpoint: "/users/${userId}/follow",
-    );
+    return await apiService.post(endpoint: "/users/${userId}/follow");
   }
 
   Future<Map<String, dynamic>> unFollowUsers(int userId) async {
-    return await apiService.delete(
-      endpoint: "/users/${userId}/follow",
-    );
+    return await apiService.delete(endpoint: "/users/${userId}/follow");
   }
+
   Future<Map<String, dynamic>> oAuthGoogleLogin(String idToken) async {
-    final response = await apiService.post(endpoint: ServerConstant.oAuthGoogleLogin, data: {'idToken' : idToken});
+    final response = await apiService.post(
+      endpoint: ServerConstant.oAuthGoogleLogin,
+      data: {'idToken': idToken},
+    );
     return response;
   }
+
   Future<Map<String, dynamic>> oAuthGithubLogin(String code) async {
-    final response = await apiService.post(endpoint: ServerConstant.oAuthExchangeCode, data: {'code' : code});
+    final response = await apiService.post(
+      endpoint: ServerConstant.oAuthExchangeCode,
+      data: {'code': code},
+    );
+    return response;
+  }
+
+  Future<Map<String, dynamic>> forgotPassword(String email) async {
+    final response = await apiService.post(
+      endpoint: ServerConstant.forgotPasswordConst,
+      data: {'email': email, 'type': "MOBILE"},
+    );
+    return response;
+  }
+
+  Future<Map<String, dynamic>> resetPassword({
+    required String password,
+    required String token,
+    required String email,
+    required int id,
+  }) async {
+    final response = await apiService.post(
+      endpoint: ServerConstant.resetPasswordForgot,
+      data: {
+        'userId': id,
+        'token': token,
+        'newPassword': password,
+        "email": email,
+      },
+    );
     return response;
   }
 }
